Nedovrsen unos datuma i status stavki u evidenciji

+++ Server/SystemOperation/EvidencijaRezSO/PromeniEvidencijaRezSO.cs
@@ -1,6 +1,8 @@
 ﻿using Common.Domain;
+using Common.Domain.Enums;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -25,16 +27,40 @@ namespace Server.SystemOperation.EvidencijaRezSO
 
             foreach (StavkaEvidencije stavka in stareStavke.Cast<StavkaEvidencije>())
             {
-                repository.Delete(stavka);
+                if (stavka.StatusStavke == StatusStavke.DODATA)
+                {
+                    stavka.Evidencija = e;
+                    repository.InsertInto(stavka);
+                    Debug.WriteLine(stavka.StatusStavke);
+                }
+                if (stavka.StatusStavke == StatusStavke.OBRISANA)
+                {
+                    repository.Delete(stavka);
+                    Debug.WriteLine(stavka.StatusStavke);
+                }
+                if (stavka.StatusStavke == StatusStavke.IZMENJENA)
+                {
+                    repository.Update(stavka);
+                    Debug.WriteLine(stavka.StatusStavke);
+                }
             }
 
-            long rb = 1;
-            foreach (StavkaEvidencije stavka in e.StavkeEvidencije)
-            {
-                stavka.Evidencija = e;
-                stavka.Rb = rb++;
-                repository.InsertInto(stavka);
-            }
+            
+
+            //foreach (StavkaEvidencije stavka in stareStavke.Cast<StavkaEvidencije>())
+            //{
+            //    repository.Delete(stavka);
+            //}
+
+            //long rb = 1;
+            //foreach (StavkaEvidencije stavka in e.StavkeEvidencije)
+            //{
+            //    stavka.Evidencija = e;
+            //    stavka.Rb = rb++;
+            //    repository.InsertInto(stavka);
+            //}
+
+
 
         }
     }
